bsort/pyproject.toml
[project]
name = "bsort"
version = "0.1.0"
description = "ML Pipeline for Object Detection"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Daffa Haj Tsaqif", email = "narutohaj00@gmail.com"},
]
dependencies = [
  "ultralytics==8.3.232",
  "requests==2.32.5",
  "opencv-python==4.12.0.88",
  "pandas==2.2.3",
  "numpy==2.2.6",
  "PyYAML==6.0.3",
  "pillow==11.3.0",
  "click>=8.3.1",
  "onnx>=1.19.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.6",
]

[project.scripts]
bsort = "src.cli:main"

[project.urls]
"Homepage" = "https://github.com/dhupee/adamata-ml-engineer-test"
"Bug Reports" = "https://github.com/dhupee/adamata-ml-engineer-test/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]

[tool.setuptools.package-dir]
bsort = "src"

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.ruff]
line-length = 70 
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
